Move track page store actions to common
