+++ src/pages/track-page/store/sagas.js
@@ -15,9 +15,9 @@ import { waitForBackendSetup } from 'store/backend/sagas'
 import { getIsReachable } from 'store/reachability/selectors'
 import { NOT_FOUND_PAGE, trackRemixesPage } from 'utils/route'
 
+import * as trackPageActions from '../../../common/store/pages/track/actions'
 import { tracksActions } from '../../../common/store/pages/track/lineup/actions'
 
-import * as trackPageActions from './actions'
 import { getTrack, getTrendingTrackRanks, getUser } from './selectors'
 
 export const TRENDING_BADGE_LIMIT = 10
